Добавил сгенерированную doxygen'ом документацию. Тупо поржать

+++ editor/lgen2diagrameditor.cpp
@@ -89,7 +89,7 @@ void LGen2DiagramEditor::deleteLink(unsigned sid, unsigned did)
     Arrow* a = 0;
     for (int i = 0; i < m_links.count(); ++i) {
         a = m_links[i];
-        if (a && (uint)a->endItem() != 0xfeeefeee  && (uint)a->startItem() != 3722304989)
+        if (a && (unsigned long long)a->endItem() != 0xfeeefeee  && (unsigned long long)a->startItem() != 3722304989)
             if (a->startItem()->id() == sid &&
                     a->endItem()->id() == did) {
                 a = m_links[i];
